+++ app/models/maths_and_physics/dqt_record.rb
@@ -9,10 +9,10 @@ module MathsAndPhysics
   #
   #   qts_award_date:     The date the teacher achieved qualified
   #                       teacher status.
-  #   itt_subject_codes:  The corresponding JAC codes to the subject
+  #   itt_subject_codes:  The corresponding JAC codes or HECOS to the subject
   #                       specialism that the teacher competed their
   #                       initial teacher training in.
-  #   degree_codes:       The corresponding JAC codes to the subject(s)
+  #   degree_codes:       The corresponding JAC codes or HECOS to the subject(s)
   #                       the teacher completed their degree in.
   class DQTRecord
     attr_reader :qts_award_date, :itt_subject_codes, :degree_codes
@@ -20,7 +20,10 @@ module MathsAndPhysics
     # Full list of JAC principal subject codes can be found:
     # https://www.hesa.ac.uk/support/documentation/jacs/jacs3-principal
     #
-    # Eligible JAC codes for this policy:
+    # Further information on the HECOS codes and their mapping against the JACS codes can be found:
+    # https://www.hesa.ac.uk/innovation/hecos
+    #
+    # Eligible JAC codes and HECOs codes for this policy:
     # https://www.gov.uk/government/publications/additional-payments-for-teaching-eligibility-and-payment-details/claim-a-payment-for-teaching-maths-or-physics-eligibility-and-payment-details#teachers-qualifications
     ELIGIBLE_JAC_CODES = [
       "G1", # Mathematics
@@ -30,16 +33,55 @@ module MathsAndPhysics
       "F3" # Physics
     ].freeze
 
+    ELIGIBLE_MATHS_HECOS_CODES = [
+      "100400", # Applied mathematics
+      "100401", # Financial mathematics
+      "100402", # Mathematical modelling
+      "100403", # Mathematics
+      "100404", # Operational research
+      "100405", # Pure mathematics
+      "100406", # Statistics
+      "101027", # Numerical analysis
+      "101028", # Engineering and industrial mathematics
+      "101029", # Computational mathematics
+      "101030", # Applied statistics
+      "101031", # Medical statistics
+      "101032", # Probability
+      "101033", # Stochastic processes
+      "101034" # Statistical modelling
+    ].freeze
+
+    ELIGIBLE_PHYSICS_HECOS_CODES = [
+      "100416", # Chemical physics
+      "100419", # Medical physics
+      "100425", # Physics
+      "100426", # Theoretical physics
+      "101060", # Applied physics
+      "101061", # Engineering physics
+      "101068", # Atmospheric physics
+      "101071", # Computational physics
+      "101074", # Radiation physics
+      "101075", # Photonics and Optical physics
+      "101076", # Laser physics
+      "101077", # Nuclear and Particle physics
+      "101223", # Condensed matter physics
+      "101300", # Quantum theory and Applications
+      "101390", # Marine physics
+      "101391" # Electromagnetism
+    ].freeze
+
     # The record transformed from a DQTReportCsv. Expected to contain the keys:
     # :qts_date              - The date the teacher achieved qualified teacher
     #                          status.
     #                          Format: %d/%m/%Y
-    # :itt_subject_jac_codes - An array of the claimants ITT subject JAC codes.
-    # :degree_jac_codes      - An array of the claimants degree JAC codes.
+    # :itt_subject_codes - An array of the claimants ITT subject JAC or HECOS codes.
+    # :degree_codes      - An array of the claimants degree JAC or HECOS codes.
+    # Previously only JAC codes were checked, however changes to the DQT mean subject codes can be either JACS or HECOS codes.
+
     def initialize(record)
       @qts_award_date = record.fetch(:qts_date)
-      @itt_subject_codes = record.fetch(:itt_subject_jac_codes)
-      @degree_codes = record.fetch(:degree_jac_codes)
+      @itt_subject_codes = record.fetch(:itt_subject_codes)
+      @degree_codes = record.fetch(:degree_codes)
     end
 
     def eligible?
@@ -57,11 +99,19 @@ module MathsAndPhysics
     end
 
     def itt_subject_maths_or_physics?
-      itt_subject_codes.any? { |jac_code| jac_code.start_with?(*ELIGIBLE_JAC_CODES) }
+      itt_subject_codes.any? { |subject_code|
+        subject_code.start_with?(*ELIGIBLE_JAC_CODES) ||
+          ELIGIBLE_MATHS_HECOS_CODES.include?(subject_code) ||
+          ELIGIBLE_PHYSICS_HECOS_CODES.include?(subject_code)
+      }
     end
 
     def maths_or_physics_degree?
-      degree_codes.any? { |jac_code| jac_code.start_with?(*ELIGIBLE_JAC_CODES) }
+      degree_codes.any? { |subject_code|
+        subject_code.start_with?(*ELIGIBLE_JAC_CODES) ||
+          ELIGIBLE_MATHS_HECOS_CODES.include?(subject_code) ||
+          ELIGIBLE_PHYSICS_HECOS_CODES.include?(subject_code)
+      }
     end
   end
 end